convertToType data object errors fixed

+++ nimble/customLearners/knn_classification.py
@@ -83,8 +83,8 @@ class KNNClassifier(CustomLearner):
         each point in trainX. The resulting matrix will be shape:
         numTestPoints x numTrainPoints.
         """
-        trainArray = self._trainX.copy('numpy array')
-        testArray = testX.copy('numpy array')
+        trainArray = self._trainX.copy('numpy array').astype(numpy.float)
+        testArray = testX.copy('numpy array').astype(numpy.float)
         # euclidean distance for each point in test
         dists = numpy.sqrt(-2 * numpy.dot(testArray, trainArray.T)
                            + numpy.sum(trainArray**2, axis=1)
